Bug fix long int

+++ src/asserts.c
@@ -777,7 +777,7 @@ void assert_not_equal_int(int expected, int result, const char *on_error_msg, co
    assert_int(expected, result, &C_TEST_FN_DESCRIPTION_ASSERT_NOT_EQ_INT, on_error_msg, on_success);
 }
 
-static void assert_longint(int expected, int result, C_TEST_FN_DESCRIPTION *desc, const char *on_error_msg, const char *on_success)
+static void assert_longint(long long int expected, long long int result, C_TEST_FN_DESCRIPTION *desc, const char *on_error_msg, const char *on_success)
 {
    static C_TEST_TYPE_LONG_INT type;
